Refactors Data in a Package
