+++ benchflow-experiment-manager/application/src/test/java/cloud/benchflow/experimentmanager/helpers/data/MinioTestData.java
@@ -1,4 +1,4 @@
-package cloud.benchflow.experimentmanager.helpers;
+package cloud.benchflow.experimentmanager.helpers.data;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
